RegretMatching, EXP3, AlphaGo Tree policies implemented and working.
Much of the complexity of Regret Matching and EXP3 has been removed from SingleTreeNode after going through the maths I realised it wasn't needed (the dangers of copying a recipe instead of understanding the principles). This means there is no need to weight be chance of choosing an action, and a single back-prop rule can be used for all of them.

I also removed the rolloutsEnabled flag on MCTSParams. This is now set simply by rolloutLength=0, which is cleaner. A new parameter for maxTreeDepth introduced at the same time. This means that we can also support Monte Carlo Search simply by setting rollout=100+, treeDepth=1.