Amended Diamant to use a new SimultaneousTurnOrder.
This moves the logic of what to do for other players who have already decided into the Player (which is where it should be).
The existing logic for doing this randomly has been moved into OSLAPlayer.
(The old code was causing problems for MCTS - as we want to use MCTS to make the decisions about what the other player do. We were getting some odd results with the old set up).

I have also amended the Heuristic to rely more on the number of gems collected - so that the aim is always to get more gems, with a bonus for being in the lead. This provides much more solid results with MCTS or OSLA in my tests.