Fix to all GameParams to use correct defaults
We need to formally _reset() after adding all tunable parameters. Otherwise we default to the value initialised in the params code...and this is overridden when we copy(). Not a problem if the same defaults are in the code and the addTunableParameter() defult, but Virus had one case where these were different.