Reduced MAX_RANGE from 10 to 8
The maximum range an attack can have in this first scenario is 9 (6 Blue + 2 Yellow + 1 Surge) for Heroes and 10 for Monsters (Goblin Archer masters have +2 from Surge), however realistically this is unlikely to occur.

For these experiments it is acceptable to tone down the maximum legal range to nudge the agents in the right direction.