Begun groundworks for Villain Heuristics
These are for the Overlord to be more in line with a traditional TTRPG Game Master, rather than an adversarial competitor.

In other words, the Overlord is to provide a challenge for the Heroes, but the Heroes will ultimately win, as the objective is to have fun, not to beat the opponents.