Fixed DescentGameState not having any listeners
James has recommended that I instead create a separate listener purely for narrative tracking, rather than combining it with the Metrics, which will enable it to print out in real time, rather than purely at the end of the game.

That will be a future step, but for now, this will service.