Make RandomPlayer work correctly, as a setup to implement MCTSPlayer
Computing available actions now actually lists an individual action for
each possible action. That is, a separate action is listed for each
possible troop that can be selected, or move that can be made, or attack
that can be executed, or command that can be applied.

However, the GUI suffers greatly from this; due to this change, the
buttons at the bottom are basically a massive list of 'select' actions,
after which they become a list of 'move' actions, etc. So, to remedy
that, I've made the updateButtons fetch a separate list of generic
actions, which rely on the highlighted cell/troop/command. I'm not sure
if circumventing the ForwardModel is wise, though...

Still to do: computing the actual maximum amount of possible moves at
one time. This is currently 10, which is too small.