Split up boilerplate action code into CQAction; optimize UI buttons
Made a lot of generic action methods work through a single abstract
CQAction class, which makes it a lot easier to check some things.

This was mainly used to improve the GUI's checking which actions are
allowed, based on the highlighted cells. This makes it possible to use
the same Action objects, without creating a massive list of possible
action buttons.

Also introduced a utility class which could be useful to move some
methods from CQGameState into, since there are several utility functions
in the CQGameState class now, which don't necessarily have to be in
there per se.