Implement A* for distance calculations, apply to where needed
Apply Floodfill to the GUI, because calculating distances requires
access to the gamestate, so the update() function would need to
pre-calculate all distances in order to make drawCell get access to the
info. Since floodfill is more efficient than doing A* on all possible
cells, that is done on every update instead.