Move utility functions to more appropriate locations
After restructuring the order of the methods in CQGameState, the utility
functions generally fit better in the list of gamestate methods, or in
another class.
Also added a verification to check if a command is already applied to a
troop, to prevent double berserking or something. That would make Winds
of Fate too overpowered (and is not part of the original game).