Add some more useful output when verbose / debugging
For parametersearch, the parallel evaluations make it confusing which
thread is doing what, when debugging is on. With these additions to
GameEvaluator, the hashCode for the evaluator that is outputting the
specific debugging line is also printed, in order to be able to
reconstruct which instance has run which order of matchups.

I did not implement the same in other debug messages, because in those
cases of parallelized messages, I feel like the information in stdout
doesn't necessarily need to be reconstructed in the same way.

I've also made the `verbose` parameter get passed on through to the
tournament after ParameterSearch, since this seems to me like expected
behaviour. If this is not desired I'll just change it back.