Reinforce .equals() and .hashCode() checks
Based on some digging in the MCTS code, reusing the tree will cause
backtracking to find an action .equal() to another action. However,
having two actions flagged as equal was fairly easy to happen
incorrectly, for example: SelectTroop on (1,1) by player 0. Even if the
troop is different, and a whole lot of other things are different, the
actions are still marked as equal.

To remedy this I just included a hash of the game state the action is
applied to, to every action. That way, equality is only met if the
actions are actually equal, and not just happen to be on the same
square.