Make game deterministic for its sole random action
I want to claim this is a pretty clever trick to ensure that applying
action A onto state S will _always_ result in the exact same state S'.
This allows me to be certain that evaluating an action on a state will
always result in the same outcome, even when that action happens to be
the one action that involves randomness (Winds of Fate).
This simply re-initializes rnd whenever getRnd is called, with the
current hashCode as a seed. This ensures the resulting number is
definitely as random as can be, but it'll also no longer be different
each time this action is evaluated.

This will help in getting a well-functioning MCTS agent to work
properly, because this will ensure that, when a winning move is found in
the tree, the full turn can certainly be executed fully in one go,
instead of having to re-evaluate the state after performing the action.