Added Lord Merick's Aftershock passive
Aftershock triggers an Attribute Test at the end of an attack made by an adjacent Hero.
Because of the way TAG handles Descent, we instead check for the attribute test in the middle of the attack - i.e. once the attacker has rolled for their attack, but before the defender rolls their defence.
This saves on computation when we get certain abilities (i.e. Splig's "Not Me!" test) which activate a test before the attack.