Added Sir Alric Farrow's Regeneration and Unmovable passives
Unmovable per the tabletop game says the Monster can choose whether it can be affected or not.
For the sake of complexity (and not constantly prompting the Overlord every time this might happen), for now Unmovable just straight-up prevents forced movement.
If there comes a point where we use ForcedMove actions where an Unmovable Monster might want to be moved, I'll change it later, but realistically most scenarios will be Heroes trying to move the Monster into disadvantageous positions.