Added Thief's Caltrops skill
This also made the Move action an IExtendedSequence, so we can interrupt it (well, 'interrupt', at the very end) to enable Caltrops.

This also opens the door for future abilities and skills which allow for Figures to react to movements.

If I ever get around to implementing Overlord Cards again properly (I probably won't) the Pit Trap or Tripwire cards would be easily implemented here.