'''Enabled''' the Bones of Woe Relic (not really)
Bones of Woe work by rolling a Blue attack die at the start of the Overlord's turns. If they get a Surge, they get to draw 1 new Overlord card into their hand.

Problem is, Overlord cards aren't implemented. So we can't do that here.

This is going to be a similar problem to what the Staff of Shadows will face, which similarly requires the Overlord Cards to function.