Fixed major error when constructing board!
This one was very important.

The original code only accounted for tiles being added beneath the starting tile when assembling (i.e. everything flowed in a bottom-rightwards direction). If any tile was assembled to the north or west of the starting tile, the bounding box would not accommodate for it, and thus cull those parts of the tiles when it came to assembly.

Now though, when we move the TopLeftCorner for the bounding box, we also adjust the size, as well as considering where the BottomRIghtCorner should be and also adjust size there.

We can now actually build other maps!